+++ app/views/payments/new.html.erb
@@ -77,7 +77,7 @@
       <button id="sweet-alert" class="submit-payment btn btn-primary">
         <div class="spinner hidden" id="spinner"></div>
         <span id="button-text">Submit payment</span>
-      </button>
+        <%= javascript_pack_tag("sweetalert") %>
       </button>
     </form>
